--- features/documentselection/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/documentselection/internalapi/DocumentSelectionEntryPoints.kt
@@ -1,6 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.documentselection.internalapi
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-interface DocumentSelectionEntryPoints
