--- features/documentselection/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/documentselection/internalapi/nav/DocumentSelectionDestinations.kt
@@ -1,9 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.documentselection.internalapi.nav
-
-import kotlinx.serialization.Serializable
-import uk.gov.onelogin.criorchestrator.libraries.navigation.NavigationDestination
-
-sealed interface DocumentSelectionDestinations : NavigationDestination {
-    @Serializable
-    data object DocumentSelection : DocumentSelectionDestinations
-}
