--- features/documentselection/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/documentselection/internalapi/nav/DocumentSelectionNavGraphProvider.kt
@@ -1,9 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.documentselection.internalapi.nav
-
-import uk.gov.onelogin.criorchestrator.libraries.navigation.NavGraphProvider
-
-/**
- * Interface which [NavGraphProvider]s must implement to contribute to the navigation graph
- * started by the 'prove your identity' feature.
- */
-interface DocumentSelectionNavGraphProvider : NavGraphProvider
