--- features/documentselection/internal-api/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/documentselection/internalapi/ExampleUnitTest.kt
@@ -1,11 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.documentselection.internalapi
-
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Test
-
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
