--- features/documentselection/internal/src/main/res/values-cy/strings.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="documentselection_passport_readmore_button">Darllenwch fwy am y mathau o ID gyda llun y gallwch eu defnyddio</string>
-</resources>
