--- features/documentselection/internal/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="documentselection_passport_readmore_button">Read more about the types of photo ID you can use</string>
-</resources>
