+++ features/select-doc/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internalapi/SelectDocumentEntryPoints.kt
@@ -1,0 +1,6 @@
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internalapi
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+interface SelectDocumentEntryPoints
