+++ features/select-doc/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internalapi/nav/SelectDocumentDestinations.kt
@@ -1,0 +1,9 @@
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internalapi.nav
+
+import kotlinx.serialization.Serializable
+import uk.gov.onelogin.criorchestrator.libraries.navigation.NavigationDestination
+
+sealed interface SelectDocumentDestinations : NavigationDestination {
+    @Serializable
+    data object SelectDocument : SelectDocumentDestinations
+}
