+++ features/select-doc/internal-api/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internalapi/ExampleUnitTest.kt
@@ -1,0 +1,11 @@
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internalapi
+
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
+
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
