+++ features/select-doc/internal/src/main/res/values-cy/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="selectdocument_passport_readmore_button">Darllenwch fwy am y mathau o ID gyda llun y gallwch eu defnyddio</string>
+</resources>
