+++ features/select-doc/internal/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="selectdocument_passport_readmore_button">Read more about the types of photo ID you can use</string>
+</resources>
