+++ features/select-doc/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/ScreenshotTest.kt
@@ -1,0 +1,26 @@
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internal
+
+import com.google.testing.junit.testparameterinjector.TestParameter
+import com.google.testing.junit.testparameterinjector.TestParameterInjector
+import org.junit.Ignore
+import org.junit.runner.RunWith
+import sergio.sastre.composable.preview.scanner.android.AndroidPreviewInfo
+import sergio.sastre.composable.preview.scanner.core.preview.ComposablePreview
+import uk.gov.onelogin.criorchestrator.libraries.screenshottesting.BaseScreenshotTest
+import uk.gov.onelogin.criorchestrator.libraries.screenshottesting.PackagePreviewsProvider
+
+private object ModulePackagePreviewsProvider : PackagePreviewsProvider()
+
+@Ignore(
+    """
+    Screenshot tests will be re-enabled when a testable screen is added to this module.
+    This is expected in https://github.com/govuk-one-login/mobile-android-cri-orchestrator/pull/183
+    """,
+)
+@RunWith(TestParameterInjector::class)
+class ScreenshotTest(
+    @TestParameter(
+        valuesProvider = ModulePackagePreviewsProvider::class,
+    )
+    preview: ComposablePreview<AndroidPreviewInfo>,
+) : BaseScreenshotTest(preview = preview)
