+++ features/select-doc/internal/src/test/snapshots/images/uk.gov.onelogin.criorchestrator.features.documentselection.internal_ScreenshotTest_screenshot[1.uk.gov.onelogin.criorchestrator.features.documentselection.internal.passport.PassportSelectionScreenKt_PreviewPassportSelectionScreen_]_dark_night.png
